refactor: implement feature-based structure
